Initial simple support for OML Config file
From [0]: “Ok here is the patch for an initial support, which will only
look for 'experiment', 'id', and 'url' fields in the passed config file.
If multiple occurrences are found, the last one will be taken.
I did not do a true XML parsing of the config file because:

    I am not sure that we want to make oml4r depend on xml parsing
libraries,
    there was some discussions to move away from XML in the config file
    there is a plan to redesign the syntax of the config file
    I am not familiar with the code in place in OML4R do deal with the
new per-stream collection URLs
”

This is probably not the best way to do it, but it is required for OMF6
examples to work [1].

[0] http://mytestbed.net/issues/1076#note-4
[1] http://mytestbed.net/issues/1076#note-8

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>