Use OML4R.logger for consistent messages
All messages are prefixed with OML4R, which might not be exactly what we
want. At the moment, the application writer needs to remember to prefix
their own application name, resulting in messages like:

  ERROR   ping-oml2: OML4R: Missing values for parameters :domain (--oml-domain) or :appName (in code)!

Not optimal, but it works.

The OML4R logger should probably be tweaked to it can take (or guess)
the name to prefix (e.g., basename of the script from which it is
called?)

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>